requirements.txt
numpy
gym
absl-py
stable_baselines3
git+git://github.com/utiasSTARS/liegroups
